requirements.txt
futures==3.1.1
gcloud==0.18.3
google-api-python-client==1.6.5
grpcio==1.10.0
grpcio-tools==1.10.0
requests==2.18.4
isodate==0.6.0
